+++ tests/test_docker_clean_old.py
@@ -1,7 +1,8 @@
 from _docker_clean_old import docker_clean_old
 import mock
 
-# uses a massive list of predefined combisons that represent possible image tags retrieved by docker
+# uses a massive list of predefined combinations that represent possible image tags retrieved by docker
+# listed image are ordered from creation date (default), which could be pretty much anything
 # according to tested flags, different results are expected to process it
 MOCK_DOCKER_LIST = [
     # bunch on intermediate images
@@ -91,6 +92,87 @@ MOCK_DOCKER_LIST = [
     "docker-registry.crim.ca/ogc/ogc-thelper-base:latest",
 ]
 
+# default args remove all 'none'
+# only 'latest' remains if other tags are also matched
+# extra tag variants are not distinguished
+DEFAULT_CALL_REMOVES = [
+    "- <none>:<none>",
+    "- <none>:<none>",
+    "- <none>:<none>",
+    "- <none>:<none>",
+    "- <none>:<none>",
+    "- devops:0.0.0",
+    "  devops:latest",
+    "- python:3.7-slim",
+    "- python:3.7-slim-buster",
+    "- python:3.7-alpine",
+    "- python:3.8-slim",
+    "  python:3.8-slim-buster",
+    "- pavics/twitcher:<none>",
+    "- pavics/twitcher:magpie-1.7.3",
+    "- pavics/twitcher:magpie-3.0.0",
+    "- pavics/twitcher:magpie-3.0.0-rc",
+    "- pavics/twitcher:magpie-3.1.0-rc",
+    "  pavics/twitcher:magpie-3.3.0-dev",
+    "- pavics/magpie:<none>",
+    "- pavics/magpie:2.0.0",
+    "- pavics/magpie:3.0.0",
+    "- pavics/magpie:3.0.0-rc",
+    "- pavics/magpie:3.1.0-rc",
+    "- pavics/magpie:3.3.0-dev",
+    "  pavics/magpie:latest",
+    "- birdhouse/twitcher:v0.5.3",
+    "- birdhouse/twitcher:v0.5.4",
+    "- birdhouse/twitcher:0.5.3",
+    "  birdhouse/twitcher:0.5.4",
+    "- video-action-recognition:<none>",
+    "- video-action-recognition:<none>",
+    "- video-action-recognition:<none>",
+    "- video-action-recognition:<none>",
+    "- video-action-recognition:<none>",
+    "- video-action-recognition:<none>",
+    "- video-action-recognition:0.3.0",
+    "- video-action-recognition:0.4.0",
+    "- video-action-recognition:0.5.0",
+    "- video-action-recognition:0.6.0",
+    "  video-action-recognition:latest",
+    "- thelper:base",
+    "  thelper:geo",
+    "- plstcharles/thelper:v0.5.0",
+    "- plstcharles/thelper:v0.6.2-geo-cuda-11.1-cudnn8-devel-ubuntu18.04",
+    "- plstcharles/thelper:v0.6.2-geo",
+    "  plstcharles/thelper:v0.6.2-geo-cuda-10.0-cudnn7-devel-ubuntu16.04",
+    "- pytorch-cuda-tester:10.0-cudnn7-devel-ubuntu16.04",
+    "  pytorch-cuda-tester:11.1-cudnn8-devel-ubuntu18.04",
+    "- nvidia/cuda:10.1-cudnn8-devel-ubuntu18.04",
+    "- nvidia/cuda:10.2-cudnn7-devel-ubuntu18.04",
+    "  nvidia/cuda:11.1-cudnn8-devel-ubuntu18.04",
+    "- lastools:0.1.0",
+    "  lastools:latest",
+    "  unidata/thredds-docker:latest",
+    "  ubuntu:20.04",
+    "  rust:latest",
+    "  registry:2",
+    "  alpine:latest",
+    "  pavics/canarieapi:0.4.3",
+    "  continuumio/miniconda3:latest",
+    "  tianon/true:latest",
+    "  stefanprodan/mgob:0.9",
+    "  mongo:3.4.0",
+    "- docker-registry.crim.ca/ogc/ogc-lidar-tb16:0.2.0",
+    "  docker-registry.crim.ca/ogc/ogc-lidar-tb16:latest",
+    "- ogc-lidar-tb16:0.2.0",
+    "  ogc-lidar-tb16:latest",
+    "- ogc-thelper-toy:0.2.1",
+    "  ogc-thelper-toy:latest",
+    "- docker-registry.crim.ca/ogc/ogc-thelper-toy:0.2.1",
+    "  docker-registry.crim.ca/ogc/ogc-thelper-toy:latest",
+    "- ogc-thelper-base:1.1.0",
+    "  ogc-thelper-base:latest",
+    "- docker-registry.crim.ca/ogc/ogc-thelper-base:1.1.0",
+    "  docker-registry.crim.ca/ogc/ogc-thelper-base:latest",
+]
+
 
 def mock_process(cmd, *_, **__):
     class MockProcess(object):
@@ -98,7 +180,7 @@ def mock_process(cmd, *_, **__):
 
         @staticmethod
         def communicate(*_, **__):
-            output = "\n".join(MOCK_DOCKER_LIST)
+            output = "\n".join("{} {}".format(*line.split(":")) for line in MOCK_DOCKER_LIST)
             return output, 0
         @property
         def stdout(self):
@@ -107,16 +189,33 @@ def mock_process(cmd, *_, **__):
             class StdOut(object):
                 @property
                 def readline(self):
-                    return ""  # just skip iterate stdout
+                    return lambda *_, **__: ""  # just skip iterate stdout
             return StdOut()
 
     return MockProcess()
 
 
+def get_log_lines(captured_logs, strip_header=False):
+    logs = [line.message for line in captured_logs.records]
+    if strip_header:
+        for i, line in enumerate(logs):
+            if line.startswith("---"):
+                return logs[i + 1:]
+    return logs
 
-def test_docker_clean_old():
+
+def test_docker_clean_old_basic():
     with mock.patch("subprocess.Popen", side_effect=mock_process) as proc:
         docker_clean_old()
-        proc.called
+        assert proc.call_count == 2, "expected 1 call to fetch images and another to remove selected ones"
+        assert proc.call_args_list[0].args[0].startswith("docker images")
+        assert proc.call_args_list[1].args[0].startswith("docker rmi")
 
 
+def test_docker_clean_old_dry_run(caplog):
+    with mock.patch("subprocess.Popen", side_effect=mock_process) as proc:
+        docker_clean_old(dry_run=True)
+        assert proc.call_count == 1, "expected 1 call to fetch images but remove images not called"
+        assert proc.call_args_list[0].args[0].startswith("docker images")
+    logs = get_log_lines(caplog, strip_header=True)
+    assert logs == DEFAULT_CALL_REMOVES